Correct temporal interpolation of GEFS swdown forcing field
zdoy was not correctly initialized to the current day of year
before calling zterp.  It was incorrectly initialized outside
the tile loop, resulting in zdoy decrementing by 1 for each pass
through the tile loop because zdoy is an intent(inout) argument
to zterp.  In my particular test, zdoy started at 2 and ended at -26859.

zdoy is now initialized inside the tile loop, right before
the call to zterp.